backend/requirements.txt
clickhouse-connect
pandas
openai
neo4j
tqdm
streamlit
pyvis
python-dotenv
requests
numpy
matplotlib
plotly